requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
openai==1.23.6
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.1.1
pytest-mock==3.14.0
